+++ R/plots_utils.R
@@ -88,6 +88,7 @@ get_date_breaks <- function(dates, threshold = 160) {
 #' @import dplyr
 #' @import ggplot2
 #' @importFrom dplyr filter
+#' @importFrom plotly ggplotly
 #' @export
 #'
 #' @examples
@@ -121,7 +122,9 @@ line_plot_data <- function(data, value, startdate = data$Date[1], title) {
   #add title if present
   if (!missing(title))
     p <- p + labs(title = title)
-  p
+  pp <- p %>%
+    ggplotly(tooltip = c("x", "y"))
+  pp
 }
 
 #' Barplot function
@@ -139,6 +142,7 @@ line_plot_data <- function(data, value, startdate = data$Date[1], title) {
 #' @import dplyr
 #' @import ggplot2
 #' @importFrom dplyr filter
+#' @importFrom plotly ggplotly
 #' @export
 #'
 #' @examples
@@ -181,6 +185,8 @@ bar_plot_data <- function(data, value, startdate = data$Date[1], rollm = FALSE, 
   #add title if present
   if (!missing(title))
     p <- p + labs(title = title)
-  p
+  pp <- p %>%
+    ggplotly(tooltip = c("x", "y"))
+  pp
 }