+++ NEWS.md
@@ -1,3 +1,4 @@
 # myfinalpkg 1.0.0 .9000
 
 * Added a `NEWS.md` file to track changes to the package.
+* Added the `plotly` package